organize dbal sorted code
