+++ src/database/redis/sorted.js
@@ -6,79 +6,10 @@ module.exports = function (redisClient, module) {
 
 	var helpers = module.helpers.redis;
 
-	module.sortedSetAdd = function (key, score, value, callback) {
-		callback = callback || function () {};
-		if (Array.isArray(score) && Array.isArray(value)) {
-			return sortedSetAddMulti(key, score, value, callback);
-		}
-		redisClient.zadd(key, score, value, function (err) {
-			callback(err);
-		});
-	};
-
-	function sortedSetAddMulti(key, scores, values, callback) {
-		if (!scores.length || !values.length) {
-			return callback();
-		}
-
-		if (scores.length !== values.length) {
-			return callback(new Error('[[error:invalid-data]]'));
-		}
-
-		var args = [key];
-
-		for(var i = 0; i < scores.length; ++i) {
-			args.push(scores[i], values[i]);
-		}
-
-		redisClient.zadd(args, function (err) {
-			callback(err);
-		});
-	}
-
-	module.sortedSetsAdd = function (keys, score, value, callback) {
-		callback = callback || function () {};
-		var multi = redisClient.multi();
-
-		for(var i = 0; i < keys.length; ++i) {
-			multi.zadd(keys[i], score, value);
-		}
-
-		multi.exec(function (err) {
-			callback(err);
-		});
-	};
-
-	module.sortedSetRemove = function (key, value, callback) {
-		callback = callback || function () {};
-		if (!value) {
-			return callback();
-		}
-		if (!Array.isArray(value)) {
-			value = [value];
-		}
-
-		helpers.multiKeyValues(redisClient, 'zrem', key, value, function (err) {
-			callback(err);
-		});
-	};
-
-	module.sortedSetsRemove = function (keys, value, callback) {
-		helpers.multiKeysValue(redisClient, 'zrem', keys, value, function (err) {
-			callback(err);
-		});
-	};
-
-	module.sortedSetsRemoveRangeByScore = function (keys, min, max, callback) {
-		callback = callback || function () {};
-		var multi = redisClient.multi();
-		for(var i = 0; i < keys.length; ++i) {
-			multi.zremrangebyscore(keys[i], min, max);
-		}
-		multi.exec(function (err) {
-			callback(err);
-		});
-	};
+	require('./sorted/add')(redisClient, module);
+	require('./sorted/remove')(redisClient, module);
+	require('./sorted/union')(redisClient, module);
+	require('./sorted/intersect')(redisClient, module);
 
 	module.getSortedSetRange = function (key, start, stop, callback) {
 		sortedSetRange('zrange', key, start, stop, false, callback);
@@ -239,62 +170,6 @@ module.exports = function (redisClient, module) {
 		multi.exec(callback);
 	};
 
-
-	module.sortedSetUnionCard = function (keys, callback) {
-		var tempSetName = 'temp_' + Date.now();
-
-		var multi = redisClient.multi();
-		multi.zunionstore([tempSetName, keys.length].concat(keys));
-		multi.zcard(tempSetName);
-		multi.del(tempSetName);
-		multi.exec(function (err, results) {
-			if (err) {
-				return callback(err);
-			}
-
-			callback(null, Array.isArray(results) && results.length ? results[1] : 0);
-		});
-	};
-
-	module.getSortedSetUnion = function (params, callback) {
-		params.method = 'zrange';
-		sortedSetUnion(params, callback);
-	};
-
-	module.getSortedSetRevUnion = function (params, callback) {
-		params.method = 'zrevrange';
-		sortedSetUnion(params, callback);
-	};
-
-	function sortedSetUnion(params, callback) {
-
-		var tempSetName = 'temp_' + Date.now();
-
-		var rangeParams = [tempSetName, params.start, params.stop];
-		if (params.withScores) {
-			rangeParams.push('WITHSCORES');
-		}
-
-		var multi = redisClient.multi();
-		multi.zunionstore([tempSetName, params.sets.length].concat(params.sets));
-		multi[params.method](rangeParams);
-		multi.del(tempSetName);
-		multi.exec(function (err, results) {
-			if (err) {
-				return callback(err);
-			}
-			if (!params.withScores) {
-				return callback(null, results ? results[1] : null);
-			}
-			results = results[1] || [];
-			var objects = [];
-			for(var i = 0; i < results.length; i += 2) {
-				objects.push({value: results[i], score: parseFloat(results[i + 1])});
-			}
-			callback(null, objects);
-		});
-	}
-
 	module.sortedSetIncrBy = function (key, increment, value, callback) {
 		redisClient.zincrby(key, increment, value, function (err, newValue) {
 			callback(err, !err ? parseFloat(newValue) : undefined);
@@ -332,11 +207,11 @@ module.exports = function (redisClient, module) {
 			maxmax = '+';
 		}
 
-		if (min !== minmin) {
-			if (!min.match(/^[\[\(]/)) min = '[' + min;
+		if (min !== minmin && !min.match(/^[\[\(]/)) {
+			min = '[' + min;
 		}
-		if (max !== maxmax) {
-			if (!max.match(/^[\[\(]/)) max = '[' + max;
+		if (max !== maxmax && !max.match(/^[\[\(]/)) {
+			max = '[' + max;
 		}
 
 		if (count) {
@@ -344,79 +219,5 @@ module.exports = function (redisClient, module) {
 		} else {
 			redisClient[method]([key, min, max], callback);
 		}
-	}
-
-	module.sortedSetIntersectCard = function (keys, callback) {
-		if (!Array.isArray(keys) || !keys.length) {
-			return callback(null, 0);
-		}
-		var tempSetName = 'temp_' + Date.now();
-
-		var interParams = [tempSetName, keys.length].concat(keys);
-
-		var multi = redisClient.multi();
-		multi.zinterstore(interParams);
-		multi.zcard(tempSetName);
-		multi.del(tempSetName);
-		multi.exec(function (err, results) {
-			if (err) {
-				return callback(err);
-			}
-
-			callback(null, results[1] || 0);
-		});
-	};
-
-	module.getSortedSetIntersect = function (params, callback) {
-		params.method = 'zrange';
-		getSortedSetRevIntersect(params, callback);
-	};
-
-	module.getSortedSetRevIntersect = function (params, callback) {
-		params.method = 'zrevrange';
-		getSortedSetRevIntersect(params, callback);
-	};
-
-	function getSortedSetRevIntersect(params, callback) {
-		var sets = params.sets;
-		var start = params.hasOwnProperty('start') ? params.start : 0;
-		var stop = params.hasOwnProperty('stop') ? params.stop : -1;
-		var weights = params.weights || [];
-
-		var tempSetName = 'temp_' + Date.now();
-
-		var interParams = [tempSetName, sets.length].concat(sets);
-		if (weights.length) {
-			interParams = interParams.concat(['WEIGHTS'].concat(weights));
-		}
-
-		if (params.aggregate) {
-			interParams = interParams.concat(['AGGREGATE', params.aggregate]);
-		}
-
-		var rangeParams = [tempSetName, start, stop];
-		if (params.withScores) {
-			rangeParams.push('WITHSCORES');
-		}
-
-		var multi = redisClient.multi();
-		multi.zinterstore(interParams);
-		multi[params.method](rangeParams);
-		multi.del(tempSetName);
-		multi.exec(function (err, results) {
-			if (err) {
-				return callback(err);
-			}
-
-			if (!params.withScores) {
-				return callback(null, results ? results[1] : null);
-			}
-			results = results[1] || [];
-			var objects = [];
-			for(var i = 0; i < results.length; i += 2) {
-				objects.push({value: results[i], score: parseFloat(results[i + 1])});
-			}
-			callback(null, objects);
-		});
 	}
 };
